Completed MVC refactoring of SudokuTile class
All JavaFX View modifications present in the SudokuTile class have been changed to interactions between the SudokuModel, PuzzleView, and PuzzleController.

The SudokuTile Class is now part of the model, only affecting the data behind the game and no longer having any direct influence on the view. It can now safely be moved to the GameLogic package as part of the model.